Verify all 80 DSA problems and add explained, tested solutions (#1)
Add independent behavioral tests for 240 C++ implementations, sanitizer checks,
problem walkthroughs, platform contract notes, and continuous verification.

+++ 01_Linear_Traversal_and_In_Place/GeeksforGeeks/GFG_Second_Largest/03_better_approach.cpp
@@ -4,17 +4,17 @@ using namespace std;
 class Solution {
 public:
     int getSecondLargest(vector<int> &arr) {
-        int largestValue = INT_MIN;
+        long long largestValue = LLONG_MIN;
         for(int i = 0; i < arr.size(); i++) {
-            largestValue = max(largestValue, arr[i]);
+            largestValue = max(largestValue, (long long)arr[i]);
         }
-        int secondLargest = INT_MIN;
+        long long secondLargest = LLONG_MIN;
         for(int i = 0; i < arr.size(); i++) {
             if(arr[i] != largestValue && arr[i] > secondLargest) {
                 secondLargest = arr[i];
             }
         }
-        if(secondLargest == INT_MIN) {
+        if(secondLargest == LLONG_MIN) {
             return -1;
         }
         return secondLargest;
